Fix model & migrations
