+++ .sequelizerc
@@ -1,8 +1,8 @@
 const path = require('path');
 
 module.exports = {
-  'config': path.resolve('src/db', 'config.js'),
-  'migrations-path': path.resolve('src/db', 'migrations'),
-  'models-path': path.resolve('src/db', 'models'),
-  'seeders-path': path.resolve('src/db', 'seeders'),
+  'config': path.resolve('src/', 'config.js'),
+  'migrations-path': path.resolve('src/', 'migrations'),
+  'models-path': path.resolve('src/', 'models'),
+  'seeders-path': path.resolve('src/', 'seeders'),
 };
