--- src/config/config.json
@@ -1,18 +1,0 @@
-{
-  "development": {
-    "username": "postgres",
-    "password": "postgres",
-    "database": "hemsut-dev",
-    "host": "127.0.0.1",
-    "port": 5432,
-    "dialect": "postgres"
-  },
-  "test": {
-    "username": "bear",
-    "password": null,
-    "database": "hemsut-test",
-    "host": "127.0.0.1",
-    "port": 5432,
-    "dialect": "postgres"
-  }
-}
