--- src/db/models/group.js
@@ -1,31 +1,0 @@
-'use strict';
-module.exports = (sequelize, DataTypes) => {
-  const Group = sequelize.define(
-    'Group',
-    {
-      id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        autoIncrement: true,
-        allowNull: false,
-      },
-
-      name: {
-        type: DataTypes.STRING,
-        allowNull: false,
-      },
-    },
-    {
-      underscored: true,
-      classMethods: {
-        associate(models) {
-          Group.belongsToMany(models.User, { through: models.UserGroup });
-          Group.hasMany(models.UserGroup);
-          Group.hasMany(models.Invitation);
-          Group.hasMany(models.Meeting);
-        },
-      },
-    },
-  );
-  return Group;
-};
