+++ src/models/group.js
@@ -1,13 +1,31 @@
+'use strict';
 module.exports = (sequelize, DataTypes) => {
-  const Group = sequelize.define('Group', {
-    name: { type: DataTypes.STRING, allowNull: false },
-  });
+  const Group = sequelize.define(
+    'Group',
+    {
+      id: {
+        type: DataTypes.INTEGER,
+        primaryKey: true,
+        autoIncrement: true,
+        allowNull: false,
+      },
 
-  Group.associate = (models) => {
-    Group.belongsToMany(models.User, { through: models.UserGroup });
-    Group.hasMany(models.UserGroup, { foreignKey: 'group_id', as: 'Members' });
-    Group.hasMany(models.Invitation, { foreignKey: 'group_id' });
-    Group.hasMany(models.Meeting, { foreignKey: 'group_id' });
-  };
+      name: {
+        type: DataTypes.STRING,
+        allowNull: false,
+      },
+    },
+    {
+      underscored: true,
+      classMethods: {
+        associate(models) {
+          Group.belongsToMany(models.User, { through: models.UserGroup });
+          Group.hasMany(models.UserGroup);
+          Group.hasMany(models.Invitation);
+          Group.hasMany(models.Meeting);
+        },
+      },
+    },
+  );
   return Group;
 };
