+++ src/models/user.js
@@ -1,26 +1,58 @@
+'use strict';
 module.exports = (sequelize, DataTypes) => {
-  const User = sequelize.define('User', {
-    username: { type: DataTypes.STRING, unique: true },
-    first_name: DataTypes.STRING,
-    last_name: DataTypes.STRING,
-    email: { type: DataTypes.STRING, unique: true },
-    indexes: [
-      {
-        unique: true,
-        fields: ['username'],
+  const User = sequelize.define(
+    'User',
+    {
+      id: {
+        type: DataTypes.INTEGER,
+        primaryKey: true,
+        autoIncrement: true,
+        allowNull: false,
       },
-      {
-        unique: true,
-        fields: ['email'],
-      },
-    ],
-  });
 
-  User.associate = (models) => {
-    User.belongsToMany(models.Group, { through: models.UserGroup });
-    User.belongsToMany(models.Meeting, { through: models.UserMeeting });
-    User.hasMany(models.UserGroup, { foreignKey: 'group_id' });
-    User.hasMany(models.UserMeeting, { foreignKey: 'meeting_id' });
-  };
+      username: {
+        type: DataTypes.STRING,
+        unique: true,
+        allowNull: false,
+      },
+
+      first_name: {
+        type: DataTypes.STRING,
+      },
+
+      last_name: {
+        type: DataTypes.STRING,
+      },
+
+      email: {
+        type: DataTypes.STRING,
+        unique: true,
+        allowNull: false,
+      },
+    },
+    {
+      underscored: true,
+      classMethods: {
+        associate(models) {
+          User.belongsToMany(models.Group, { through: models.UserGroup });
+          User.belongsToMany(models.Meeting, { through: models.UserMeeting });
+          User.hasMany(models.UserGroup);
+          User.hasMany(models.UserMeeting);
+        },
+      },
+    },
+    {
+      indexes: [
+        {
+          unique: true,
+          fields: ['username'],
+        },
+        {
+          unique: true,
+          fields: ['email'],
+        },
+      ],
+    },
+  );
   return User;
 };
